wego/demand-back-end-repo/requirements.txt
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
coverage==7.5.0
decorator==5.1.1
Django==4.1.13
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.6.1
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mysqlclient==2.2.4
numpy==1.26.4
openapi-codec==1.3.2
orjson==3.9.15
packaging==23.2
PyJWT==2.8.0
pymongo==3.12.3
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
setuptools==69.2.0
shortuuid==1.0.12
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
sqlparse==0.2.4
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
